(feat): implemented tap detaction
