+++ lib/constants/constants.dart
@@ -1,0 +1,6 @@
+  import 'package:flutter/material.dart';
+
+final BorderRadius innerBorderRadius = BorderRadius.circular(50);
+  const  BorderRadius outterBorderRadius =  BorderRadius.all(Radius.circular(50));
+ const double compnentWidth = 350;
+  const  double compnentHeight = 100.0;
