--- lib/controllers/painter_controller.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/rendering.dart';
-
-class SumdgePainter extends CustomPainter{
-  @override
-  void paint(Canvas canvas, Size size) {
-    // TODO: implement paint
-  }
-
-  @override
-  bool shouldRepaint(covariant CustomPainter oldDelegate) {
-    // TODO: implement shouldRepaint
-    throw UnimplementedError();
-  }
-
-}
