+++ lib/utils/gesturedetector/gesturedetector_utils.dart
@@ -1,0 +1,25 @@
+import 'dart:async';
+import 'dart:typed_data';
+import 'dart:ui' as ui;
+
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
+
+
+// void onPanStart(DragStartDetails details) {
+//   print('User started drawing');
+//   final box = context.findRenderObject() as RenderBox;
+//   final point = box.globalToLocal(details.globalPosition);
+//   print(point);
+// }
+
+// void onPanUpdate(DragUpdateDetails details) {
+//   final box = context.findRenderObject() as RenderBox;
+//   final point = box.globalToLocal(details.globalPosition);
+//   print(point);
+// }
+
+// void onPanEnd(DragEndDetails details) {
+//   print('User ended drawing');
+// }
